bpf: Allow to use 24 bits for security identities
This changes the BPF datapath to use 32 bit integers to handle security
identities. It also adjusts the skb mark handling to split up the identity and
use part of the existing magic marker to store the new extended 8 bits of the
identity.

This commit does not extend the the allocator to use additional identity space
yet.

A single 16 bit identity field remained in the lxcmap. It was only used for
debugging and its value was questionable as the endpoint ID to identity mapping
is available elsewhere. In the interest of reduced complexity, usage was
removed and the field is marked unused.

+++ bpf/bpf_lxc.c
@@ -94,7 +94,7 @@ static inline int ipv6_l3_from_lxc(struct __sk_buff *skb,
 	struct ct_state ct_state = {};
 	void *data, *data_end;
 	union v6addr *daddr, orig_dip;
-	uint16_t dstID = WORLD_ID;
+	uint32_t dstID = WORLD_ID;
 	__u32 tunnel_endpoint = 0;
 
 	if (unlikely(!is_valid_lxc_src_mac(eth)))
@@ -420,7 +420,7 @@ static inline int handle_ipv4_from_lxc(struct __sk_buff *skb)
 	struct ct_state ct_state_new = {};
 	struct ct_state ct_state = {};
 	__be32 orig_dip;
-	uint16_t dstID = WORLD_ID;
+	uint32_t dstID = WORLD_ID;
 	__u32 tunnel_endpoint = 0;
 
 	if (!revalidate_data(skb, &data, &data_end, &ip4))
